% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{.find_extdata_dir}
\alias{.find_extdata_dir}
\title{Trouver le repertoire extdata}
\usage{
.find_extdata_dir()
}
\value{
Chemin vers le repertoire extdata.
}
\description{
Cherche le repertoire \code{inst/extdata} dans le package installe ou dans
l'arborescence source (mode developpement).
}
\keyword{internal}
