% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{.find_python_dir}
\alias{.find_python_dir}
\title{Trouver le repertoire des modules Python}
\usage{
.find_python_dir()
}
\value{
Chemin vers le repertoire contenant les modules Python.
}
\description{
Cherche le repertoire \code{inst/python} dans le package installe ou dans
l'arborescence source (mode developpement).
}
\keyword{internal}
