% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{ensure_python_deps}
\alias{ensure_python_deps}
\title{Installer les dependances Python pour pplmatch}
\usage{
ensure_python_deps(envname = "pplmatch")
}
\arguments{
\item{envname}{Nom de l'environnement virtuel (defaut : "pplmatch").
Le virtualenv est cree dans \code{~/.virtualenvs/}.}
}
\value{
\code{TRUE} (invisible) si l'installation a reussi.
}
\description{
Cree un environnement virtuel Python nomme "pplmatch" et y installe le
package \code{rapidfuzz}, necessaire pour l'appariement fuzzy dans
\code{\link{pplmatchQC}}.
}
\details{
Cette fonction ne doit etre appelee qu'**une seule fois** par machine.
Apres l'installation, le package active automatiquement le virtualenv au
chargement.


Prerequis : Python 3 (>= 3.8) doit etre installe sur le systeme.
Le virtualenv est gere par \code{reticulate} et se trouve dans
\code{~/.virtualenvs/pplmatch/}.
}
\examples{
\dontrun{
# Premiere utilisation du package :
ensure_python_deps()

# Ensuite, pplmatchQC() fonctionne directement :
result <- pplmatchQC(corpus, members)
}

}
\seealso{
\code{\link{pplmatchQC}} pour la fonction de matching principale.
}
