% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data.R
\name{qc_members}
\alias{qc_members}
\title{Charger la table des deputes historiques du Quebec}
\usage{
qc_members()
}
\value{
Un tibble avec les colonnes : \code{full_name}, \code{party_id},
  \code{gender}, \code{legislature_id}, \code{other_names}, \code{district_id}.
}
\description{
Lit le fichier CSV des deputes integre au package et retourne un tibble
pret a etre utilise avec \code{\link{pplmatchQC}}.
}
\examples{
\dontrun{
members <- qc_members()
head(members)
}

}
