% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mandates.R
\name{qc_persons}
\alias{qc_persons}
\title{Charger la table des personnes}
\usage{
qc_persons()
}
\value{
Un tibble : \code{person_id}, \code{full_name}, \code{other_names},
  \code{assnat_url}.
}
\description{
Charger la table des personnes
}
